Revert "Merge pull request #795 from mitchellmarkoff/db-6550-next-wp-search-input"
This reverts commit c6a15b791bdc89c8b9e4082e6cd7228a8407148b, reversing
changes made to 9ffd79d1783852ae25719fc9c4f9a214ecd7476e.
