--- .changeset/good-apples-wonder.md
@@ -1,6 +1,0 @@
----
-"create-pantheon-decoupled-kit": minor
----
-
-[cli] - Add generator for a next-wp-search-api-add-on
-[next-wp-search-api-add-on] - Add's <SearchInput /> component to the layout and header of the next-wp starter.
