+++ .vscode/extensions.json
@@ -3,7 +3,6 @@
     "esbenp.prettier-vscode",
     "dbaeumer.vscode-eslint",
     "streetsidesoftware.code-spell-checker",
-    "redhat.vscode-yaml",
-    "bierner.comment-tagged-templates"
+    "redhat.vscode-yaml"
   ]
 }
