+++ packages/create-pantheon-decoupled-kit/src/generators/index.ts
@@ -1,13 +1,11 @@
-import { gatsbyWpAcfAddon } from './gatsby-wp-acf-addon.generator';
-import { gatsbyWp } from './gatsby-wp.generator';
-import { nextDrupalSearchApiAddon } from './next-drupal-search-api-addon.generator';
-import { nextDrupalUmamiAddon } from './next-drupal-umami-addon.generator';
-import { nextDrupal } from './next-drupal.generator';
-import { nextWpAcfAddon } from './next-wp-acf-addon.generator';
-import { nextWpSearchApiAddon } from './next-wp-search-api-addon.generator';
 import { nextWp } from './next-wp.generator';
+import { nextDrupal } from './next-drupal.generator';
+import { gatsbyWp } from './gatsby-wp.generator';
+import { nextDrupalUmamiAddon } from './next-drupal-umami-addon.generator';
+import { nextWpAcfAddon } from './next-wp-acf-addon.generator';
+import { gatsbyWpAcfAddon } from './gatsby-wp-acf-addon.generator';
 import { tailwindcssAddon } from './tailwindcss-addon.generator';
-
+import { nextDrupalSearchApiAddon } from './next-drupal-search-api-addon.generator';
 
 export const decoupledKitGenerators = [
 	nextWp,
@@ -18,5 +16,4 @@ export const decoupledKitGenerators = [
 	gatsbyWpAcfAddon,
 	tailwindcssAddon,
 	nextDrupalSearchApiAddon,
-	nextWpSearchApiAddon,
 ];
