+++ packages/create-pantheon-decoupled-kit/src/index.ts
@@ -2,12 +2,11 @@ import chalk from 'chalk';
 import inquirer, { QuestionCollection } from 'inquirer';
 import minimist, { Opts as MinimistOptions, ParsedArgs } from 'minimist';
 import {
-	Input,
+	type DecoupledKitGenerator,
+	type TemplateData,
 	isDrupalCms,
 	isString,
 	isWpCms,
-	type DecoupledKitGenerator,
-	type TemplateData,
 } from './types';
 import { actionRunner, getHandlebarsInstance, helpMenu } from './utils/index';
 
@@ -189,7 +188,7 @@ To see this list at any time, use the --help command.`;
 	const actionsResult = await actionRunner({
 		actions,
 		templateData,
-		data: args as Input,
+		data: args,
 		handlebars: hbs,
 	});
 	args?.silent || console.log(chalk.blueBright(actionsResult));
