+++ packages/create-pantheon-decoupled-kit/src/templates/next-drupal-search-api-addon/components/layout.jsx.hbs
@@ -1,0 +1,1 @@
+{{> nextDrupalLayout}}
