--- packages/create-pantheon-decoupled-kit/src/templates/next-drupal-search-api-addon/components/layout.jsx.ts
@@ -1,10 +1,0 @@
-import { TemplateFn } from '@cli/types';
-import { layoutTemplate } from '@partials/nextjs-shared/layout';
-
-const jsx: TemplateFn = ({ data }) =>
-	/* jsx */ `${layoutTemplate({
-		search: data.search,
-		cmsType: data.cmsType,
-	})}`;
-
-export default jsx;
