--- packages/create-pantheon-decoupled-kit/src/templates/next-drupal-search-api-addon/components/search-input.jsx.ts
@@ -1,6 +1,0 @@
-import { TemplateFn } from '@cli/types';
-import { searchInputTemplate } from '@partials/nextjs-shared/searchInput';
-
-const jsx: TemplateFn = () => `${searchInputTemplate()}`;
-
-export default jsx;
