--- packages/create-pantheon-decoupled-kit/src/templates/next-wp-search-api-addon/components/searchInput.module.css.ts
@@ -1,6 +1,0 @@
-import { TemplateFn } from '@cli/types';
-import { searchInputCSSModules } from '@partials/nextjs-shared/searchInputCSSModules';
-
-const css: TemplateFn = () => `${searchInputCSSModules()}`;
-
-export default css;
