+++ packages/create-pantheon-decoupled-kit/src/templates/partials/nextjs-shared/nextjsLayoutCSS.hbs
@@ -1,6 +1,4 @@
-import { taggedTemplateHelpers as utils } from '@cli/utils';
-
-export const layoutCSSModules = (search: boolean) => /* css */ `.layout {
+.layout {
 	display: flex;
 	flex-direction: column;
 	min-height: 100dvh;
@@ -34,14 +32,12 @@ export const layoutCSSModules = (search: boolean) => /* css */ `.layout {
 	color: var(--blue-100);
 }
 
-${utils.if(
-	search,
-	/* css */ `.searchHeaderContainer {
-    display: flex; 
-    margin: var(--x-auto);
-    flex-direction: row; 
-    flex-wrap: wrap; 
-    justify-content: center; 
-}`,
-)}
-`;
+{{#if search}}
+.searchHeaderContainer {
+	display: flex; 
+	margin: var(--x-auto);
+	flex-direction: row; 
+	flex-wrap: wrap; 
+	justify-content: center; 
+}
+{{/if}}
