+++ packages/create-pantheon-decoupled-kit/src/templates/tailwindless-next/components/layout.module.css.hbs
@@ -1,0 +1,1 @@
+{{> nextjsLayoutCSS}}
