--- packages/create-pantheon-decoupled-kit/src/templates/tailwindless-next/components/layout.module.css.ts
@@ -1,7 +1,0 @@
-import { TemplateFn } from '@cli/types';
-import { layoutCSSModules } from '@partials/nextjs-shared/layoutCSSModules';
-
-const css: TemplateFn = ({ data }) =>
-	/* css */ `${layoutCSSModules(data.search)}`;
-
-export default css;
