+++ packages/create-pantheon-decoupled-kit/src/types.ts
@@ -30,10 +30,6 @@ export interface GatsbyWPData extends BaseGeneratorData {
 	gatsby: true;
 }
 
-export interface SearchApiData extends BaseGeneratorData {
-	search: boolean;
-}
-
 /**
  * Generators need prompts to get user data not provided by CLI arguments
  */
@@ -102,8 +98,7 @@ export type Action = (config: ActionConfig) => Promise<string> | string;
 export type ActionRunner = (config: ActionRunnerConfig) => Promise<string>;
 
 type InputIndex = BaseGeneratorData &
-	GatsbyWPData &
-	SearchApiData & {
+	GatsbyWPData & {
 		_: string[];
 		appName: string;
 		outDir: string;
@@ -114,7 +109,6 @@ type InputIndex = BaseGeneratorData &
 		force: boolean;
 		silent: boolean;
 		tailwindcss: boolean;
-		cmsType: CMSType['Drupal'] | CMSType['WordPress'];
 	};
 
 /**
